fix: remove circle import
