+++ src/types.d.ts
@@ -562,3 +562,5 @@ export interface IRetroItem {
   updatedAt?: string;
 }
 /** ******************************************************* */
+
+export type AvatarEditPanel = 'MAIN' | 'CROPPER' | 'COLLECTION';
